requirements.txt
alembic==1.6.5
anyio==3.4.0
appdirs==1.4.3
apturl==0.5.2
asgiref==3.4.1
attrs==21.2.0
backports.zoneinfo==0.2.1
bcrypt==3.1.7
beautifulsoup4==4.10.0
blinker==1.4
Brlapi==0.7.0
certifi==2019.11.28
chardet==3.0.4
cliapp==1.20180812.1
click==8.0.1
cmdtest==0.32+git
colorama==0.4.3
command-not-found==0.3
cryptography==2.8
cupshelpers==1.0
dbus-python==1.2.16
defer==1.0.6
distlib==0.3.0
distro==1.4.0
distro-info===0.23ubuntu1
Django==4.0.1
django-etc==1.3.1
django-hitcount==1.3.5
django-resized==0.3.11
djangorestframework==3.12.4
dnspython==2.1.0
duplicity==0.8.12.0
email-validator==1.1.3
entrypoints==0.3
fastapi==0.70.0
fastapi-jwt-auth==0.5.0
fastapi-utils==0.2.1
fasteners==0.14.1
filelock==3.0.12
Flask==2.0.1
Flask-Login==0.5.0
Flask-Migrate==3.0.1
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.15.1
future==0.18.2
gpg===1.13.1-unknown
greenlet==1.1.0
h11==0.12.0
httplib2==0.14.0
httptools==0.2.0
idna==2.8
importlib-metadata==1.5.0
iniconfig==1.1.1
itsdangerous==2.0.1
Jinja2==3.0.1
kazam==1.4.5
keyring==18.0.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lockfile==0.12.2
louis==3.12.0
macaroonbakery==1.3.1
Mako==1.1.0
Markdown==3.1.1
MarkupSafe==2.0.1
migrations==0.1.1
monotonic==1.5
more-itertools==4.2.0
MouseInfo==0.1.3
mysqlclient==2.0.3
netifaces==0.10.4
oauthlib==3.1.0
olefile==0.46
packaging==20.3
paramiko==2.6.0
peewee==3.14.8
pexpect==4.6.0
Pillow==9.0.0
pluggy==1.0.0
protobuf==3.6.1
psycopg2-binary==2.9.3
py==1.11.0
PyAutoGUI==0.9.53
pycairo==1.16.2
pycrypto==2.6.1
pycups==1.9.73
pydantic==1.8.2
PyGetWindow==0.0.9
Pygments==2.3.1
PyGObject==3.36.0
pyinotify==0.9.6
PyJWT==1.7.1
pymacaroons==0.13.0
PyMsgBox==1.0.9
PyMySQL==1.0.2
PyNaCl==1.3.0
pyOpenSSL==19.0.0
pyparsing==2.4.6
pyperclip==1.8.2
PyRect==0.1.4
pyRFC3339==1.1
PyScreeze==0.1.28
pytest==6.2.5
python-apt==2.0.0+ubuntu0.20.4.3
python-dateutil==2.7.3
python-debian===0.1.36ubuntu1
python-dotenv==0.19.2
python-editor==1.0.4
python3-xlib==0.15
pytweening==1.0.4
pytz==2021.3
pywhatkit==5.2
pyxdg==0.26
PyYAML==5.3.1
reportlab==3.5.34
requests==2.22.0
requests-unixsocket==0.2.0
SecretStorage==2.3.1
simplejson==3.16.0
six==1.14.0
sniffio==1.2.0
soupsieve==2.3.1
SQLAlchemy==1.4.17
sqlparse==0.4.2
starlette==0.16.0
systemd-python==234
toml==0.10.2
ttystatus==0.38
typing-extensions==4.0.1
ubuntu-advantage-tools==20.3
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.25.8
usb-creator==0.3.7
uvicorn==0.15.0
uvloop==0.16.0
virtualenv==20.0.17
wadllib==1.3.3
watchgod==0.7
websockets==10.1
Werkzeug==2.0.1
wikipedia==1.4.0
WTForms==2.3.3
xkit==0.0.0
zipp==1.0.0
